# Improve interop performance against dictionary