I'm adding the test case and also the option which shows converting to native `JsValue` objects from JSON deserialization. If use case is reading then native types should produce best performance. I might have made mistakes so take it with a grain of salt.

The delegate building is now cached against AST node reference so it should align quite well with the idea of preparing and caching AST when needed.

## Jint.Benchmark.InteropLambdaBenchmark

| **Diff**|Type|Method|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |ClrObject|InlineCSharp|13.958 μs|0.1449 μs|30.08 KB|
| **New** |	|	| **12.110 μs (-13%)** | **0.2307 μs** | **28.2 KB (-6%)** |
| Old |ClrObject|InlineEngineInvoke|14.468 μs|0.1819 μs|30 KB|
| **New** |	|	| **12.620 μs (-13%)** | **0.1171 μs** | **28.13 KB (-6%)** |
| Old |ClrObject|Inline|14.810 μs|0.1906 μs|29.61 KB|
| **New** |	|	| **12.871 μs (-13%)** | **0.1735 μs** | **27.73 KB (-6%)** |
| Old |ClrObject|ForLoopEngineInvoke|20.211 μs|0.1940 μs|34.92 KB|
| **New** |	|	| **18.729 μs (-7%)** | **0.3184 μs** | **32.66 KB (-6%)** |
| Old |ClrObject|ForLoop|20.554 μs|0.3373 μs|34.53 KB|
| **New** |	|	| **18.925 μs (-8%)** | **0.3124 μs** | **33.05 KB (-4%)** |
| Old |JsonNode|ForLoop|49.571 μs|0.7526 μs|60.63 KB|
| **New** |	|	| **34.749 μs (-30%)** | **0.6410 μs** | **47.97 KB (-21%)** |
| Old |JsonNode|ForLoopEngineInvoke|50.650 μs|0.6915 μs|61.02 KB|
| **New** |	|	| **34.759 μs (-31%)** | **0.6888 μs** | **47.58 KB (-22%)** |
| Old |JsonNode|Inline|56.960 μs|1.1315 μs|75.87 KB|
| **New** |	|	| **41.791 μs (-27%)** | **0.8285 μs** | **62.82 KB (-17%)** |
| Old |JsonNode|InlineCSharp|61.993 μs|1.2329 μs|76.34 KB|
| **New** |	|	| **41.995 μs (-32%)** | **0.8274 μs** | **63.29 KB (-17%)** |
| Old |JsonNode|InlineEngineInvoke|62.023 μs|1.2133 μs|76.26 KB|
| **New** |	|	| **42.142 μs (-32%)** | **0.8274 μs** | **63.21 KB (-17%)** |
| Old |Dictionary|ForLoop|37.053 μs|0.3455 μs|47.11 KB|
| **New** |	|	| **22.510 μs (-39%)** | **0.2309 μs** | **42.58 KB (-10%)** |
| Old |Dictionary|ForLoopEngineInvoke|37.151 μs|0.4640 μs|47.5 KB|
| **New** |	|	| **23.255 μs (-37%)** | **0.2261 μs** | **33.13 KB (-30%)** |
| Old |Dictionary|Inline|2,421.450 μs|31.9202 μs|150.02 KB|
| **New** |	|	| **23.602 μs (-99%)** | **0.4647 μs** | **42.5 KB (-72%)** |
| Old |Dictionary|InlineEngineInvoke|2,426.210 μs|42.1312 μs|150.42 KB|
| **New** |	|	| **23.608 μs (-99%)** | **0.3058 μs** | **42.11 KB (-72%)** |
| Old |Dictionary|InlineCSharp|2,439.785 μs|47.8304 μs|150.5 KB|
| **New** |	|	| **23.728 μs (-99%)** | **0.3679 μs** | **33.52 KB (-78%)** |
| Old |JsValue|InlineCSharp|8.711 μs|0.1299 μs|18.91 KB|
| **New** |	|	| **8.943 μs (+3%)** | **0.1251 μs** | **18.91 KB (0%)** |
| Old |JsValue|InlineEngineInvoke|9.205 μs|0.1786 μs|18.83 KB|
| **New** |	|	| **9.063 μs (-2%)** | **0.0782 μs** | **18.83 KB (0%)** |
| Old |JsValue|Inline|9.278 μs|0.1288 μs|18.44 KB|
| **New** |	|	| **9.175 μs (-1%)** | **0.1086 μs** | **18.44 KB (0%)** |
| Old |JsValue|ForLoopEngineInvoke|14.978 μs|0.1502 μs|23.75 KB|
| **New** |	|	| **14.975 μs (0%)** | **0.1385 μs** | **23.75 KB (0%)** |
| Old |JsValue|ForLoop|15.345 μs|0.2494 μs|23.36 KB|
| **New** |	|	| **15.260 μs (-1%)** | **0.1876 μs** | **23.36 KB (0%)** |


fixes #2087