@lahma 2025-04-03
> I want to see the issue updated with the new results now.

I copied the benchmarks and modified to support running as single (params for scenario) with addition of projecting to JsValue data (ideal). So I would guess this reflects the original problem.