@lahma 2025-04-03
The codebase is mostly benchmarked against JS scenarios so these interop ones can be tricky. You could probably start a consulting business about the performance traps with interop and JS...