@lahma 2025-04-04
Here's the full non-compared benchmark result:

```

BenchmarkDotNet v0.14.0, Windows 11 (10.0.26100.3624)
12th Gen Intel Core i9-12900H, 1 CPU, 20 logical and 14 physical cores
.NET SDK 9.0.202
  [Host]     : .NET 9.0.3 (9.0.325.11113), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.3 (9.0.325.11113), X64 RyuJIT AVX2


```
| Method              | Type       | Mean      | Error     | StdDev    | Ratio | RatioSD | Rank | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------- |----------- |----------:|----------:|----------:|------:|--------:|-----:|-------:|-------:|----------:|------------:|
| InlineCSharp        | ClrObject  | 12.110 μs | 0.2307 μs | 0.2265 μs |  0.65 |    0.02 |    1 | 2.2888 | 0.0153 |   28.2 KB |        0.86 |
| InlineEngineInvoke  | ClrObject  | 12.620 μs | 0.1171 μs | 0.1096 μs |  0.67 |    0.01 |    1 | 2.2888 | 0.0153 |  28.13 KB |        0.86 |
| Inline              | ClrObject  | 12.871 μs | 0.1735 μs | 0.1449 μs |  0.69 |    0.01 |    1 | 2.2583 | 0.0153 |  27.73 KB |        0.85 |
| ForLoop             | ClrObject  | 18.729 μs | 0.3184 μs | 0.2978 μs |  1.00 |    0.02 |    2 | 2.6550 |      - |  32.66 KB |        1.00 |
| ForLoopEngineInvoke | ClrObject  | 18.925 μs | 0.3124 μs | 0.2922 μs |  1.01 |    0.02 |    2 | 2.6855 |      - |  33.05 KB |        1.01 |
|                     |            |           |           |           |       |         |      |        |        |           |             |
| ForLoopEngineInvoke | JsonNode   | 34.749 μs | 0.6410 μs | 0.5996 μs |  1.00 |    0.02 |    1 | 3.9063 |      - |  47.97 KB |        1.01 |
| ForLoop             | JsonNode   | 34.759 μs | 0.6888 μs | 0.6443 μs |  1.00 |    0.03 |    1 | 3.8452 |      - |  47.58 KB |        1.00 |
| Inline              | JsonNode   | 41.791 μs | 0.8285 μs | 1.2899 μs |  1.20 |    0.04 |    2 | 4.8828 |      - |  62.82 KB |        1.32 |
| InlineCSharp        | JsonNode   | 41.995 μs | 0.8274 μs | 0.7740 μs |  1.21 |    0.03 |    2 | 4.8828 |      - |  63.29 KB |        1.33 |
| InlineEngineInvoke  | JsonNode   | 42.142 μs | 0.8274 μs | 1.2128 μs |  1.21 |    0.04 |    2 | 4.8828 |      - |  63.21 KB |        1.33 |
|                     |            |           |           |           |       |         |      |        |        |           |             |
| InlineCSharp        | Dictionary | 22.510 μs | 0.2309 μs | 0.2047 μs |  0.97 |    0.01 |    1 | 3.4485 | 0.0305 |  42.58 KB |        1.29 |
| ForLoop             | Dictionary | 23.255 μs | 0.2261 μs | 0.2005 μs |  1.00 |    0.01 |    1 | 2.6855 |      - |  33.13 KB |        1.00 |
| InlineEngineInvoke  | Dictionary | 23.602 μs | 0.4647 μs | 0.4120 μs |  1.01 |    0.02 |    1 | 3.4485 | 0.0305 |   42.5 KB |        1.28 |
| Inline              | Dictionary | 23.608 μs | 0.3058 μs | 0.2861 μs |  1.02 |    0.01 |    1 | 3.4180 | 0.0305 |  42.11 KB |        1.27 |
| ForLoopEngineInvoke | Dictionary | 23.728 μs | 0.3679 μs | 0.3441 μs |  1.02 |    0.02 |    1 | 2.7161 |      - |  33.52 KB |        1.01 |
|                     |            |           |           |           |       |         |      |        |        |           |             |
| InlineCSharp        | JsValue    |  8.943 μs | 0.1251 μs | 0.1045 μs |  0.59 |    0.01 |    1 | 1.5411 |      - |  18.91 KB |        0.81 |
| InlineEngineInvoke  | JsValue    |  9.063 μs | 0.0782 μs | 0.0694 μs |  0.59 |    0.01 |    1 | 1.5259 |      - |  18.83 KB |        0.81 |
| Inline              | JsValue    |  9.175 μs | 0.1086 μs | 0.0963 μs |  0.60 |    0.01 |    1 | 1.4954 |      - |  18.44 KB |        0.79 |
| ForLoopEngineInvoke | JsValue    | 14.975 μs | 0.1385 μs | 0.1296 μs |  0.98 |    0.01 |    2 | 1.9379 |      - |  23.75 KB |        1.02 |
| ForLoop             | JsValue    | 15.260 μs | 0.1876 μs | 0.1755 μs |  1.00 |    0.02 |    2 | 1.8921 |      - |  23.36 KB |        1.00 |